On tap Java

+++ src/DataTypes/Ex3ChuyenDoi.java
@@ -13,7 +13,7 @@ public class Ex3ChuyenDoi {
     public static void convertMimutesIntoOffDay(double min){
         double mimutes = 60 *24*365;
         int day = (int)(min/60/24)% 360;
-        double year = (double)(min/mimutes);
+        long year = (long) (min/mimutes);
         System.out.println((int)min + " phut  " +"," + day + " ngay  " + ","+ year + "  nam ");
     }
 }
